requirements.txt
flask==2.3.3
gunicorn==22.0.0
pyjwt==2.8.0
bcrypt==4.0.1
flask-cors==4.0.0
python-dotenv==1.0.0
marshmallow==3.20.1
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-limiter==3.5.0
pytest==7.4.0
pytest-flask==1.2.0